+++ packages/ai/src/agent/agent.ts
@@ -2,11 +2,13 @@ import { GenerateTextResult } from '../generate-text/generate-text-result';
 import { StreamTextResult } from '../generate-text/stream-text-result';
 import { ToolSet } from '../generate-text/tool-set';
 import { Prompt } from '../prompt/prompt';
-import { InferUITools, UIMessage } from '../ui/ui-messages';
 
 /**
- * An agent is a reusable component that that has tools and that
- * can generate or stream content.
+ * An Agent receives a prompt (text or messages) and generates or streams an output
+ * that consists of steps, tool calls, data parts, etc.
+ *
+ * You can implement your own Agent by implementing the `Agent` interface,
+ * or use the `ToolLoopAgent` class.
  */
 export interface Agent<
   TOOLS extends ToolSet = {},
@@ -14,14 +16,20 @@ export interface Agent<
   OUTPUT_PARTIAL = never,
 > {
   /**
+   * The specification version of the agent interface. This will enable
+   * us to evolve the agent interface and retain backwards compatibility.
+   */
+  readonly version: 'agent-v1';
+
+  /**
    * The id of the agent.
    */
-  id: string | undefined;
+  readonly id: string | undefined;
 
   /**
    * The tools that the agent can use.
    */
-  tools: TOOLS;
+  readonly tools: TOOLS;
 
   /**
    * Generates an output from the agent (non-streaming).
@@ -32,11 +40,4 @@ export interface Agent<
    * Streams an output from the agent (streaming).
    */
   stream(options: Prompt): StreamTextResult<TOOLS, OUTPUT_PARTIAL>;
-
-  /**
-   * Creates a response object that streams UI messages to the client.
-   */
-  respond(options: {
-    messages: UIMessage<never, never, InferUITools<TOOLS>>[];
-  }): Response;
 }