+++ packages/ai/src/agent/tool-loop-agent.ts
@@ -5,29 +5,35 @@ import { streamText } from '../generate-text/stream-text';
 import { StreamTextResult } from '../generate-text/stream-text-result';
 import { ToolSet } from '../generate-text/tool-set';
 import { Prompt } from '../prompt/prompt';
-import { convertToModelMessages } from '../ui/convert-to-model-messages';
-import { InferUITools, UIMessage } from '../ui/ui-messages';
-import { BasicAgentSettings } from './basic-agent-settings';
 import { Agent } from './agent';
+import { ToolLoopAgentSettings } from './tool-loop-agent-settings';
 
 /**
- * The Agent class provides a structured way to encapsulate LLM configuration, tools,
- * and behavior into reusable components.
+ * A tool loop agent is an agent that runs tools in a loop. In each step,
+ * it calls the LLM, and if there are tool calls, it executes the tools
+ * and calls the LLM again in a new step with the tool results.
  *
- * It handles the agent loop for you, allowing the LLM to call tools multiple times in
- * sequence to accomplish complex tasks.
- *
- * Define agents once and use them across your application.
+ * The loop continues until:
+ * - A finish reasoning other than tool-calls is returned, or
+ * - A tool that is invoked does not have an execute function, or
+ * - A tool call needs approval, or
+ * - A stop condition is met (default stop condition is stepCountIs(20))
  */
-export class BasicAgent<
+export class ToolLoopAgent<
   TOOLS extends ToolSet = {},
   OUTPUT = never,
   OUTPUT_PARTIAL = never,
 > implements Agent<TOOLS, OUTPUT, OUTPUT_PARTIAL>
 {
-  private readonly settings: BasicAgentSettings<TOOLS, OUTPUT, OUTPUT_PARTIAL>;
+  readonly version = 'agent-v1';
 
-  constructor(settings: BasicAgentSettings<TOOLS, OUTPUT, OUTPUT_PARTIAL>) {
+  private readonly settings: ToolLoopAgentSettings<
+    TOOLS,
+    OUTPUT,
+    OUTPUT_PARTIAL
+  >;
+
+  constructor(settings: ToolLoopAgentSettings<TOOLS, OUTPUT, OUTPUT_PARTIAL>) {
     this.settings = settings;
   }
 
@@ -65,18 +71,5 @@ export class BasicAgent<
       stopWhen: this.settings.stopWhen ?? stepCountIs(20),
       ...options,
     });
-  }
-
-  /**
-   * Creates a response object that streams UI messages to the client.
-   */
-  respond(options: {
-    messages: UIMessage<never, never, InferUITools<TOOLS>>[];
-  }): Response {
-    return this.stream({
-      prompt: convertToModelMessages(options.messages, { tools: this.tools }),
-    }).toUIMessageStreamResponse<
-      UIMessage<never, never, InferUITools<TOOLS>>
-    >();
   }
 }